rm condvar, add simple wait-check loop
as the executor is not multithreaded, the condvar isn't doing what we
think it's doing, and is in fact just blocking the whole executor.

in lieu of a hook/channel/flag to determine the state of a given
request/response, we are just using the fact that storage credentials
are erased as the determining factor that the request was processed
correctly.

the 0..10 loop of 50 milliseconds equates to the original timeout
duration but also gives more opportunities (via the timeout's .await)
for the server and fake_server to make progress