[RSDK-9365] fix flaky test (#372)
as the executor is not multithreaded, using a sync condvar would just block the whole executor. 

in lieu of a hook/channel/flag to determine the state of a given request/response, we are using the fact that storage credentials are erased as the flag that the request was processed correctly.

the await-loop gives more opportunities
for the server and fake_server to make progress